+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* Add the method `assertDiagnostic(...)` that accepts a compilation, a diagnostic and a message contents to the `AbstractProcessorTest` class. Also added `assertErrorDiagnostic(...)` and `assertWarningDiagnostic(...)` helper methods to make testing the most common diagnostic error messages easy. Deprecated the existing `assertDiagnosticPresent(...)` method as it's functionality has been subsumed by the new methods.
+
 ### [v0.42](https://github.com/realityforge/proton/tree/v0.42) (2020-03-18) · [Full Changelog](https://github.com/realityforge/proton/compare/v0.41...v0.42)
 
 Changes in this release: